Milestone2/requirements.txt
pandas
matplotlib
scipy
seaborn
numpy
scikit-learn
requests
beautifulsoup4
psycopg2-binary
SQLAlchemy
docker
kafka-python

